All conflict resolved

+++ src/gui/KilometerConverter.java
@@ -14,8 +14,8 @@ import javax.swing.JTextField;
 public class KilometerConverter {
 
 	public static void main(String[] args) {
-		KilometerConverterWindow gui=new KilometerConverterWindow("Kilometer Converter");
-		gui.initializeComponents(); 
+		KilometerConverterWindow gui = new KilometerConverterWindow("Kilometer Converter");
+		gui.initializeComponents();
 	}
 }
 
@@ -24,31 +24,29 @@ class KilometerConverterWindow extends JFrame {
 	private JTextField kiloTextField;
 	private JButton calcButton;
 	private JPanel panel;
-	
+
 	public KilometerConverterWindow(String title) {
 		super(title);
 	}
 
-	private class CalcButtonListener implements ActionListener{
+	private class CalcButtonListener implements ActionListener {
 
 		@Override
 		public void actionPerformed(ActionEvent e) {
-			double input=Integer.parseInt(kiloTextField.getText());
-			double mile=input*0.621371;
-			JOptionPane.showMessageDialog(null, input+ " KM in mile = "+mile+" miles");
-			
+			double input = Double.parseDouble(kiloTextField.getText());
+			double mile = input * 0.621371;
+			JOptionPane.showMessageDialog(null, input + " KM in mile = " + mile + " miles");
+
 		}
-		
+
 	}
-	
-	
-	
+
 	public void initializeComponents() {
 
 		messageLabel = new JLabel("Enter a distance in kilometers");
 		kiloTextField = new JTextField(10);
 		calcButton = new JButton("Calculate");
-		CalcButtonListener list=new CalcButtonListener();
+		CalcButtonListener list = new CalcButtonListener();
 		calcButton.addActionListener(list);
 		panel = new JPanel();
 		panel.add(messageLabel);
@@ -57,6 +55,6 @@ class KilometerConverterWindow extends JFrame {
 		add(panel);
 		setSize(400, 200);
 		setVisible(true);
-		
+
 	}
 }
